# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ft_transcendence API
  description: >
    API REST du backend ft_transcendence (Fastify v5 + TypeScript).
    Auth JWT (access token Bearer 15 min) + refresh token en cookie httpOnly
    (SameSite=Strict). Le cookie est posé Path=/auth par le backend et réécrit en
    Path=/api/auth par le proxy Vite pour le navigateur (invariant I4). Chat temps réel
    hors-OpenAPI via WebSocket `/ws/chat` (documenté en fin de fichier pour mémoire).
  version: 0.9.0
  contact:
    name: ft_transcendence team
servers:
  # Origine applicative PRINCIPALE (navigateur) : tout passe par le proxy Vite same-origin.
  # Les chemins ci-dessous sont relatifs à cette base — /auth/login s'appelle donc
  # https://localhost:5173/api/auth/login côté navigateur.
  - url: https://localhost:5173/api
    description: App proxifiée (origine unique navigateur, I4)
  # Backend en direct : diagnostic, tests e2e Python et curl uniquement (hors navigateur app).
  - url: https://localhost:3000
    description: Backend direct (diagnostic / tests, HTTPS certificat auto-signé)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token signé (payload `{ sub }`). Les tempTokens 2FA
        (`{ sub, pending: 'totp' }`) sont refusés sur les routes protégées.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
      required: [error]
    ValidationError:
      type: object
      description: Erreur de validation Zod (issues brutes).
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: too_small
              path:
                type: array
                items:
                  type: string
              message:
                type: string
                example: au moins 8 caractères
      required: [errors]
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required: [ok]

    # ---------- User ----------
...
    BlockEntry:
      type: object
      description: Un user bloqué (retourné par GET /blocks).
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        since:
          type: string
          format: date-time
          description: createdAt de la ligne `blocks`.
      required: [id, pseudo, since]

    # ---------- Catalog (games + ladders) ----------
...
  responses:
    Unauthorized:
      description: Token absent, invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /blocks:
    get:
      tags: [Blocks]
      summary: Liste des users que j'ai bloqués
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des blocages
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlockEntry'
                required: [blocks]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /blocks/{userId}:
    post:
      tags: [Blocks]
      summary: Bloquer un user
      description: >
        Bloque l'user et rompt toute amitié existante entre les deux
        (DELETE friendship dans les deux sens). Pas idempotent : un 2e
        appel renvoie 400 'already blocked'.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Blocage effectué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: >
            Param :userId non-uuid (ValidationError) ; auto-blocage tenté ou déjà bloqué
            — SQLSTATE 23505 (Error)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User cible introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Blocks]
      summary: Débloquer un user
      description: >
        Retire la ligne `blocks` de moi vers cet user. Idempotent
        (200 même si pas bloqué).
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Déblocage effectué (ou aucun blocage existant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: Param :userId non-uuid (ValidationError)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # =====================================================================
  # GAMES
  # =====================================================================
...